+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### 2.15.3
 
+- Fix it should reconnect pbx on notification if it has been in background for more than 10s
 - Embed:
   - Fix it should emit call_update correctly